requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
django-filter==24.3
django-storages==1.14.4
boto3==1.35.76
Pillow==11.0.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
gunicorn==23.0.0
whitenoise==6.8.2
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
factory-boy==3.3.1
Faker==33.1.0
drf-spectacular==0.28.0
redis==5.2.1
celery==5.4.0
